Refactor modules directory layout
