+++ nuclear-engagement/inc/Modules/TOC/includes/class-nuclen-toc-headings.php
@@ -7,8 +7,10 @@
 
 declare(strict_types=1);
 
+namespace NuclearEngagement\Modules\TOC;
+
 if ( ! defined( 'ABSPATH' ) ) {
-	exit;
+    exit;
 }
 
 /**
