+++ nuclear-engagement/inc/Modules/TOC/includes/class-nuclen-toc-utils.php
@@ -12,8 +12,10 @@
 
 declare(strict_types=1);
 
+namespace NuclearEngagement\Modules\TOC;
+
 if ( ! defined( 'ABSPATH' ) ) {
-	exit;
+    exit;
 }
 
 /**
