+++ nuclear-engagement/inc/Modules/TOC/includes/class-nuclen-toc-view.php
@@ -9,8 +9,10 @@
 
 declare(strict_types=1);
 
+namespace NuclearEngagement\Modules\TOC;
+
 if ( ! defined( 'ABSPATH' ) ) {
-	exit;
+    exit;
 }
 
 use NuclearEngagement\SettingsRepository;
