+++ nuclear-engagement/includes/AssetVersions.php
@@ -52,10 +52,10 @@ final class AssetVersions {
 			'front_js'         => NUCLEN_PLUGIN_DIR . 'front/js/nuclen-front.js',
 			'theme_bright_css' => NUCLEN_PLUGIN_DIR . 'front/css/nuclen-theme-bright.css',
 			'theme_dark_css'   => NUCLEN_PLUGIN_DIR . 'front/css/nuclen-theme-dark.css',
-			'toc_admin_css'    => NUCLEN_PLUGIN_DIR . 'modules/toc/assets/css/nuclen-toc-admin.css',
-			'toc_admin_js'     => NUCLEN_PLUGIN_DIR . 'modules/toc/assets/js/nuclen-toc-admin.js',
-			'toc_front_css'    => NUCLEN_PLUGIN_DIR . 'modules/toc/assets/css/nuclen-toc-front.css',
-			'toc_front_js'     => NUCLEN_PLUGIN_DIR . 'modules/toc/assets/js/nuclen-toc-front.js',
+                        'toc_admin_css'    => NUCLEN_PLUGIN_DIR . 'inc/Modules/TOC/assets/css/nuclen-toc-admin.css',
+                        'toc_admin_js'     => NUCLEN_PLUGIN_DIR . 'inc/Modules/TOC/assets/js/nuclen-toc-admin.js',
+                        'toc_front_css'    => NUCLEN_PLUGIN_DIR . 'inc/Modules/TOC/assets/css/nuclen-toc-front.css',
+                        'toc_front_js'     => NUCLEN_PLUGIN_DIR . 'inc/Modules/TOC/assets/js/nuclen-toc-front.js',
 		);
 
 		$versions = array();
