+++ nuclear-engagement/includes/Plugin.php
@@ -81,10 +81,10 @@ class Plugin {
 		OptinData::init();
 
                 // TOC module
-                require_once plugin_dir_path( __FILE__ ) . '../modules/toc/loader.php';
+                require_once plugin_dir_path( __FILE__ ) . '../inc/Modules/TOC/loader.php';
 
                 // Summary module
-                require_once plugin_dir_path( __FILE__ ) . '../modules/summary/loader.php';
+                require_once plugin_dir_path( __FILE__ ) . '../inc/Modules/Summary/loader.php';
 
 		$this->loader = new Loader();
 	}
